# openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: KesselResourceService API
    version: 0.0.1
paths:
    /api/v1/resources:
        put:
            tags:
                - KesselResourceService
            operationId: KesselResourceService_UpdateResource
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateResourceRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateResourceResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - KesselResourceService
            operationId: KesselResourceService_CreateResource
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateResourceRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateResourceResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        delete:
            tags:
                - KesselResourceService
            operationId: KesselResourceService_DeleteResource
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DeleteResourceRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResourceResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        CreateResourceRequest:
            type: object
            properties:
                rhel_host:
                    $ref: '#/components/schemas/RhelHost'
                k8s_cluster:
                    $ref: '#/components/schemas/K8sCluster'
        CreateResourceResponse:
            type: object
            properties: {}
        DeleteResourceRequest:
            type: object
            properties:
                rhel_host:
                    $ref: '#/components/schemas/RhelHost'
                k8s_cluster:
                    $ref: '#/components/schemas/K8sCluster'
        DeleteResourceResponse:
            type: object
            properties: {}
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        K8sCluster:
            type: object
            properties:
                metadata:
                    $ref: '#/components/schemas/Metadata'
                resourceData:
                    $ref: '#/components/schemas/K8sClusterResourceData'
        K8sClusterResourceData:
            type: object
            properties:
                external_cluster_id:
                    type: string
                kube_version:
                    type: string
        Metadata:
            type: object
            properties:
                id:
                    readOnly: true
                    type: string
                    description: Kessel Asset Inventory generated identifier.
                created_at:
                    readOnly: true
                    type: string
                    description: Date and time when the inventory item was first reported.
                    format: date-time
                updated_at:
                    readOnly: true
                    type: string
                    description: Date and time when the inventory item was last updated.
                    format: date-time
                deleted_at:
                    readOnly: true
                    type: string
                    description: Date and time when the inventory item was deleted.
                    format: date-time
        RhelHost:
            type: object
            properties:
                metadata:
                    $ref: '#/components/schemas/Metadata'
                resourceData:
                    $ref: '#/components/schemas/RhelHostResourceData'
        RhelHostResourceData:
            type: object
            properties:
                host_id:
                    type: string
                os_version:
                    type: string
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        UpdateResourceRequest:
            type: object
            properties:
                rhel_host:
                    $ref: '#/components/schemas/RhelHost'
                k8s_cluster:
                    $ref: '#/components/schemas/K8sCluster'
        UpdateResourceResponse:
            type: object
            properties: {}
tags:
    - name: KesselResourceService